refactor(checktranslations): run script as js file
now ts-node library is no need anymore because the script is compiled to javascript
