+++ .gitignore
@@ -1,4 +1,5 @@
 build
+bin
 # Logs
 logs
 *.log
